vim-patch:8.1.1960: fold code is spread out
Problem:    Fold code is spread out.
Solution:   Move fold functions to fold.c.
https://github.com/vim/vim/commit/db022f3ffb57c5ff4b97b4e93a994d11e4c10466
